remove unused content files, updates to some logic on settings
